move time-tree into lib folder
