+++ lib/time-tree.js
